chore: extract addHexPrefix to shared/lib/add-hex-prefix
addHexPrefix is a pure utility used by UI, shared, and background.
Moving it to shared lets UI (and shared/lib/swaps-utils.js) import
it without crossing the architectural boundary;
app/scripts/lib/util keeps re-exporting it for existing background
callers.

The two multi-symbol files (ui/selectors/selectors.js,
ui/store/actions.ts) that import addHexPrefix alongside
getEnvironmentType are intentionally left for a follow-up to avoid
conflicts with the parallel getEnvironmentType PR (#43024).
