+++ shared/lib/add-hex-prefix.ts
@@ -1,0 +1,21 @@
+/**
+ * Prefixes a hex string with '0x' or '-0x' and returns it. Idempotent.
+ *
+ * @param str - The string to prefix.
+ * @returns The prefixed string.
+ */
+export const addHexPrefix = (str: string) => {
+  if (typeof str !== 'string' || str.match(/^-?0x/u)) {
+    return str;
+  }
+
+  if (str.match(/^-?0X/u)) {
+    return str.replace('0X', '0x');
+  }
+
+  if (str.startsWith('-')) {
+    return str.replace('-', '-0x');
+  }
+
+  return `0x${str}`;
+};
