+++ shared/lib/swaps-utils.js
@@ -14,9 +14,7 @@ import {
   TOKEN_API_BASE_URL,
 } from '../constants/swaps';
 import { SECOND } from '../constants/time';
-// TODO: Remove restricted import
-// eslint-disable-next-line import-x/no-restricted-paths
-import { addHexPrefix } from '../../app/scripts/lib/util';
+import { addHexPrefix } from './add-hex-prefix';
 import { isValidHexAddress } from './hexstring-utils';
 import { isEqualCaseInsensitive } from './string-utils';
 import { decimalToHex } from './conversion.utils';
